+++ lib/mega.dart
@@ -1,5 +1,8 @@
 library mega;
 
-export 'package:mega/src/colors/mega_colors.dart';
+export 'package:mega/src/styles/mega_styles.dart';
 export 'package:mega/src/widgets/mega_icon_button.dart';
 export 'package:mega/src/widgets/mega_drawer_item.dart';
+export 'package:mega/src/widgets/mega_header_bar.dart';
+export 'package:mega/src/widgets/mega_linked_buttons.dart';
+export 'package:mega/src/widgets/mega_linked_icon_button.dart';