+++ lib/src/widgets/mega_drawer_item.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:mega/src/colors/mega_colors.dart';
+import 'package:mega/src/styles/mega_styles.dart';
 import 'package:mega/src/utils/theme_picker.dart';
 
 class MegaDrawerItem extends StatefulWidget {
@@ -36,58 +36,58 @@ class _MegaDrawerItemState extends State<MegaDrawerItem> {
   Widget build(BuildContext context) {
     if (widget.onPressed != null) disabled = false;
     return Padding(
-      padding: const EdgeInsets.all(3.0),
+      padding: MegaStyle.small,
       child: Stack(
         alignment: Alignment.centerLeft,
         children: [
           InkWell(
             onTap: widget.onPressed,
             onHover: (hover) => setState(() => hovering = hover),
-            borderRadius: const BorderRadius.all(Radius.circular(5.0)),
+            borderRadius: MegaStyle.borderRadius,
             focusNode: widget.focusNode,
             autofocus: widget.autofocus ?? false,
             child: AnimatedContainer(
-              height: 34,
-              duration: const Duration(milliseconds: 200),
-              curve: Curves.ease,
+              height: MegaStyle.buttonSize,
+              duration: MegaStyle.basicDuration,
+              curve: MegaStyle.basicCurve,
               decoration: BoxDecoration(
-                borderRadius: const BorderRadius.all(Radius.circular(5.0)),
+                borderRadius: MegaStyle.borderRadius,
                 color: disabled
                     ? ThemePicker.of(context).pick(
-                        light: MegaColors.backgroundColorLight,
-                        dark: MegaColors.backgroundColorDark,
+                        light: MegaStyle.backgroundColorLight,
+                        dark: MegaStyle.backgroundColorDark,
                       )
                     : widget.isActive
                         ? ThemePicker.of(context).pick(
-                            light: MegaColors.hoverColorLight,
-                            dark: MegaColors.hoverColorDark,
+                            light: MegaStyle.hoverColorLight,
+                            dark: MegaStyle.hoverColorDark,
                           )
                         : hovering
                             ? ThemePicker.of(context).pick(
-                                light: MegaColors.hoverColorLight,
-                                dark: MegaColors.hoverColorDark,
+                                light: MegaStyle.hoverColorLight,
+                                dark: MegaStyle.hoverColorDark,
                               )
                             : ThemePicker.of(context).pick(
-                                light: MegaColors.backgroundColorLight,
-                                dark: MegaColors.backgroundColorDark,
+                                light: MegaStyle.backgroundColorLight,
+                                dark: MegaStyle.backgroundColorDark,
                               ),
               ),
-              padding: const EdgeInsets.only(right: 8.0),
+              padding: MegaStyle.right,
               child: Row(
                 children: [
                   SizedBox(
-                    width: 34,
+                    width: MegaStyle.buttonSize,
                     child: Icon(
                       widget.icon,
-                      size: 17,
+                      size: MegaStyle.iconSize,
                       color: disabled
                           ? ThemePicker.of(context).pick(
-                              light: MegaColors.iconColorDisabledLight,
-                              dark: MegaColors.iconColorDisabledDark,
+                              light: MegaStyle.iconColorDisabledLight,
+                              dark: MegaStyle.iconColorDisabledDark,
                             )
                           : ThemePicker.of(context).pick(
-                              light: MegaColors.iconColorLight,
-                              dark: MegaColors.iconColorDark,
+                              light: MegaStyle.iconColorLight,
+                              dark: MegaStyle.iconColorDark,
                             ),
                     ),
                   ),
@@ -99,15 +99,13 @@ class _MegaDrawerItemState extends State<MegaDrawerItem> {
             ),
           ),
           AnimatedContainer(
-            height: 17,
-            width: 3,
-            duration: const Duration(milliseconds: 200),
-            curve: Curves.ease,
+            height: widget.isActive ? MegaStyle.iconSize : 0,
+            width: MegaStyle.halfPadding,
+            duration: MegaStyle.basicDuration,
+            curve: MegaStyle.basicCurve,
             decoration: BoxDecoration(
-              borderRadius: const BorderRadius.all(Radius.circular(5.0)),
-              color: widget.isActive
-                      ? widget.accentColor
-                      : Colors.transparent,
+              borderRadius: MegaStyle.borderRadius,
+              color: widget.accentColor,
             ),
           ),
         ],